+++ src/breakpoint.rs
@@ -239,9 +239,8 @@ graphql_object!(SBBreakpoint: super::debugger::SBDebugger | &self | {
         self.is_valid()
     }
 
-    // TODO(bm) This should be i32
-    field id() -> i64 {
-        self.id() as i64
+    field id() -> i32 {
+        self.id()
     }
 
     field is_enabled() -> bool {
@@ -261,13 +260,13 @@ graphql_object!(SBBreakpoint: super::debugger::SBDebugger | &self | {
     }
 
     // TODO(bm) This should be u32
-    field ignore_count() -> i64 {
-        self.ignore_count() as i64
+    field ignore_count() -> i32 {
+        self.ignore_count() as i32
     }
 
     // TODO(bm) This should be u32
-    field hit_count() -> i64 {
-        self.hit_count() as i64
+    field hit_count() -> i32 {
+        self.hit_count() as i32
     }
 
     // TODO(bm) Make this work. (Lifetimes.)
